requirements.txt
Django==3.2
django-environ==0.4.5
django-filter==2.4.0
django-graphql-geojson==0.1.4
django-graphql-jwt==0.3.1
django-loginas==0.3.9
django-safedelete==1.0.0
djangorestframework==3.12.4
djangorestframework-jwt==1.11.0
docutils==0.17.1
fastdiff==0.2.0
future==0.18.2
graphene==2.1.8
graphql-core<3
graphql-relay<3
graphene-django>=2.15.0
Rx<2
html5lib==1.1
idna<3
importlib-metadata==4.0.1
inflection==0.5.1
iso8601==0.1.14
Jinja2==2.11.3
wheel==0.36.2
twine==3.4.1
jsonfield==3.1.0
keyring==23.0.1
MarkupSafe==1.1.1
more-itertools==8.7.0
numpy==1.20.2
packaging==20.9
pandas==1.2.4
Pillow==8.2.0
pkginfo==1.7.0
pluggy==0.13.1
promise==2.3
psycopg2-binary==2.8.6
py==1.10.0
pycparser==2.20
Pygments==2.8.1
PyJWT<=2.0.0
pyparsing==2.4.7
Pyphen==0.10.0
pyramda==0.1
pytest==6.2.3
pytest-cov==2.11.1
pytest-django==4.2.0
python-dateutil==2.8.1
pytz==2021.1
readme-renderer==29.0
requests==2.25.1
requests-toolbelt==0.9.1
rescape-graphene==0.4.7
rescape-python-helpers==0.0.42
simplejson==3.17.2
singledispatch==3.6.1
six==1.15.0
snapshottest==0.6.0
sqlparse==0.4.1
termcolor==1.1.0
tinycss2==1.1.0
tqdm==4.60.0
twine==3.4.1
typing==3.7.4.3
urllib3==1.26.4
utils==1.0.1
wasmer==1.0.0
wcwidth==0.2.5
WeasyPrint==52.5
webencodings==0.5.1
zipp==3.4.1
django-reversion==3.0.9
django-reversion-compare==0.14.0
pur==5.4.0
